Fix incorrect location to stylesheet
The URI being requested from the browser was including a file system path. This fixes the URI to show the proper path relative to the plugin directory. It also adds a version number parameter to the URI for cache busting in the same manner that it is used in the AbstractGateway class.